Bug que acontecia no processo de converter menos unários. Resolvido fazendo o limite do for ser atualizado toda vez

+++ main.cpp
@@ -8,7 +8,7 @@ using namespace std;
 int main()
 {
     // Recebe número de casos
-    int cases;
+    size_t cases;
     cin >> cases;
     cin.ignore();
